Fixed router port whitelist after save reload
Currently the bIsPortWhiteList boolean is not saved to the save file. This causes all routers to enter the port blacklist mode after a reload of the save game, which causes previously open ports to become closed. This patch adds the same SaveGame property to the boolean which the bIsAddrWhitelist (which is not affected by this issue) also has.