Add Wireless Network support with dedicated buildable
Introduced the Wireless Access Point, unlocked with Network Router, which can be connected to a Radar Tower in order to receive and transmit messages over the air.

Messages could be transmitted only to towers which are connected to a WAP and are in range. The range is calculated using the same Formula from SF for scanning, so an higher tower as an higher range.